Fixed some Windows compile issues

+++ c_cpp/goldsrc/player_movement/src/pm_debug.c
@@ -72,9 +72,9 @@ GOLDSRC_PLAYERMOVEMENT_PUBLIC(void) PM_ShowClipBox(void)
 		pmove->player_maxs[pmove->usehull],
 		org,
 		pmove->server ? 132 : 0,
-		0.1);
+		0.1f);
 
-	PM_ParticleLine(org, org, pmove->server ? 132 : 0, 0.1, 5.0);
+	PM_ParticleLine(org, org, pmove->server ? 132 : 0, 0.1f, 5.0f);
 /*
 	{
 		int i;
@@ -117,7 +117,7 @@ GOLDSRC_PLAYERMOVEMENT_PUBLIC(void) PM_ParticleLine(pmvec3_t start, pmvec3_t end
 		for ( i = 0; i < 3; i++ )
 			curpos[i] = start[i] + curdist * diff[i];
 
-		pmove->PM_Particle(curpos, pcolor, life, 0, vert);
+		pmove->PM_Particle(curpos, pcolor, life, 0, (int)vert);
 		curdist += linestep;
 	}
 }
@@ -288,7 +288,7 @@ GOLDSRC_PLAYERMOVEMENT_PUBLIC(void) PM_ViewEntity(void)
 
 	VectorCopy(pmove->origin, origin);
 
-	fup = 0.5 * (pmove->player_mins[pmove->usehull][2] + pmove->player_maxs[pmove->usehull][2]);
+	fup = 0.5f * (pmove->player_mins[pmove->usehull][2] + pmove->player_maxs[pmove->usehull][2]);
 	fup += pmove->view_ofs[2];
 	fup -= 4;
